Modify: use trait
